refactor: rename projects-viewmodels to project-viewmodels
