+++ server/src/features/projects/types/index.ts
@@ -1,6 +1,6 @@
 export * from "./project-data-service.type";
 export * from "./project-filter.type";
 export * from "./project-tag-data-service.type";
-export * from "./projects-cache.interface";
 export * from "./project-tags-cache.interface";
-export * from "./projects-viewmodels.type";
+export * from "./project-viewmodels.type";
+export * from "./projects-cache.interface";
